Changed filenames in test to names without _test. Added makefile for test. Fixed a lot of tests. Fixed closestpair (use of new geometry/indexed). Fixed gcc3 problems with float/double ambiguities in geometry.
bk: 3dd2835dLDMsoFIRJpX2Gi6xAHbrLQ

+++ geometry/point_ops.cpp
@@ -15,7 +15,7 @@ template <class P> typename P::coord_type dz(P p, P q) { return q.z - p.z; }
 
 #include <cmath>
 template <class P> double dist(P p) { return sqrt(dist2(p)); }
-template <class P> double angle(P p) { return atan2(p.y, p.x); }
+template <class P> double angle(P p) { return atan2((double)p.y, (double)p.x);}
 template <class P> double theta(P p) {
   double x = p.x, y = p.y; if (x==0 && y==0) return 0;
   double t = y / (x<0^y<0 ? x-y : x+y);
